Cleanup test classes for ide warnings and re-organize junit tests
